pdf2csv1/requirements.txt
Flask==2.3.3
tabula-py==2.8.2
pandas==2.1.1
Werkzeug==2.3.7
jpype1==1.4.1
gunicorn==21.2.0
openjdk==11.0.2
